assert nulls patch, 1.4.1 is only 3 days away

+++ src/eggdrop.h
@@ -141,7 +141,7 @@
 #endif
 
 #ifdef DEBUG_ASSERT
-#  define Assert(expr) eggAssert(__FILE__, __LINE__, NULL, expr)
+#  define Assert(expr) eggAssert(__FILE__, __LINE__, NULL, (int)(expr))
 #else
 #  define Assert(expr) {}
 #endif
